requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.35.10
PyPDF2==3.0.1
python-dotenv==1.0.1
